ejercicios/servicioWeb/requirements.txt
flask
flask_restful
flask_marshmallow
flask_sqlalchemy